# openapi_spec.yaml
openapi: 3.0.0
info:
  title: EDEN System API
  description: AI awakening system with Gmail integration, cybersecurity monitoring, and dual AI personas
  version: 1.0.0
  contact:
    name: EDEN System
    email: spxcemxrmxid@gmail.com
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://eden-sanctuary-production.up.railway.app
    description: Production server (if deployed)
security:
  - oauth2: []
paths:
...
  /api/gmail/messages:
    get:
      summary: Get Gmail Messages
      description: Retrieve messages from Gmail with optional query filtering
      operationId: getGmailMessages
      parameters:
        - name: query
          in: query
          description: Gmail search query (e.g., 'is:unread', 'from:example@gmail.com')
          schema:
            type: string
          required: false
        - name: max_results
          in: query
          description: Maximum number of messages to return
          schema:
            type: integer
            default: 10
            maximum: 50
          required: false
      responses:
        '200':
          description: List of Gmail messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  count:
                    type: integer
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        subject:
                          type: string
                        from:
                          type: string
                        to:
                          type: string
                        date:
                          type: string
                        snippet:
                          type: string
                        body:
                          type: string
  /api/gmail/messages/search:
    post:
      summary: Search Gmail Messages
      description: Search messages with advanced Gmail query syntax
      operationId: searchGmailMessages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Gmail search query
                  example: "is:unread from:boss@company.com"
                max_results:
                  type: integer
                  description: Maximum results to return
                  default: 50
              required:
                - query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  count:
                    type: integer
                  messages:
                    type: array
                    items:
                      type: object
  /api/gmail/messages/send:
    post:
      summary: Send Email
      description: Send an email via Gmail
      operationId: sendGmailMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                  description: Recipient email address
                  example: "recipient@example.com"
                subject:
                  type: string
                  description: Email subject
                  example: "Hello from EDEN"
                body:
                  type: string
                  description: Email body text
                  example: "This is a test email."
              required:
                - to
                - subject
                - body
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  sent:
                    type: boolean
                  message_id:
                    type: string
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            gmail.read: Read Gmail messages
            gmail.send: Send Gmail messages
            system.read: Read system status and metrics
            system.write: Control system parameters
...